Add files via upload
A few graphs on here - you have to put each unzipped FARS data file in your downloads (or edit the part that pulls in the files), theres a few graphs with some weird functions and dataframes to make them work correctly. Could probably be optimized but it doesn't seem to lag the program at all.